key-vault/nested-item: moving the wayward func
